# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Campus RPG API 문서
  version: 0.1.0
  description: |
    현재 백엔드에 구현된 API를 기준으로 작성한 OpenAPI 초안입니다.
    `backend/server.js`와 각 라우트 모듈의 실제 경로를 반영합니다.
servers:
  - url: http://localhost:8888
    description: 로컬 백엔드 서버
...
paths:
...
  /api/items:
    get:
      tags: [shop]
      summary: 상점 아이템 목록 조회
      description: 아이템 ID 오름차순으로 상점 카탈로그를 반환합니다.
      responses:
        "200":
          description: 아이템 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
        "500":
          description: 아이템 목록 조회 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
  /api/wallet:
    get:
      tags: [wallet]
      security:
        - bearerAuth: []
      summary: 사용자 코인 잔액 조회
      description: 인증된 사용자의 현재 코인 잔액을 반환합니다.
      responses:
        "200":
          description: 코인 잔액
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletResponse"
        "400":
          description: 잘못된 사용자 ID입니다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
        "401":
          description: Bearer 토큰이 없거나 유효하지 않습니다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
        "403":
          description: 요청 사용자와 토큰 사용자가 일치하지 않습니다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
        "404":
          description: 사용자를 찾을 수 없습니다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
        "500":
          description: 잔액 조회 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
  /api/inventory:
    get:
      tags: [inventory]
      security:
        - bearerAuth: []
      summary: 사용자 인벤토리 조회
      description: 인증된 사용자의 인벤토리 목록을 반환합니다.
      responses:
        "200":
          description: 인벤토리 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
        "400":
          description: 잘못된 사용자 ID입니다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
        "401":
          description: Bearer 토큰이 없거나 유효하지 않습니다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
        "403":
          description: 요청 사용자와 토큰 사용자가 일치하지 않습니다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
        "404":
          description: 사용자를 찾을 수 없습니다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
        "500":
          description: 인벤토리 조회 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
  /api/inventory/purchase:
    post:
      tags: [inventory]
      security:
        - bearerAuth: []
      summary: 아이템 구매
      description: |
        인증된 사용자 코인에서 아이템 가격을 차감하고 인벤토리에 아이템을 추가합니다.
        요청 본문에 userId를 함께 보내면 토큰 사용자와 일치하는지 확인합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseRequest"
      responses:
        "200":
          description: 구매 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseResponse"
        "400":
          description: 요청 본문이 올바르지 않습니다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
        "401":
          description: Bearer 토큰이 없거나 유효하지 않습니다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
        "402":
          description: 코인이 부족합니다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsufficientFundsResponse"
        "403":
          description: 요청 사용자와 토큰 사용자가 일치하지 않습니다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
        "404":
          description: 아이템 또는 사용자를 찾을 수 없습니다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
        "500":
          description: 구매 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCodeResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorCodeResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: INVALID_BODY
...
    Item:
      type: object
      required: [id, name, description, price, imageUrl, iconEmoji, effectType]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Coffee
        description:
          type: string
          example: Restores focus for studying.
        price:
          type: integer
          example: 30
        imageUrl:
          type: string
          nullable: true
          example: /images/items/coffee.png
        iconEmoji:
          type: string
          nullable: true
          example: "\u2615"
        effectType:
          type: string
          nullable: true
          example: focus
    InventoryItem:
      type: object
      required:
        [
          itemId,
          quantity,
          name,
          description,
          price,
          imageUrl,
          iconEmoji,
          effectType,
        ]
      properties:
        itemId:
          type: integer
          example: 1
        quantity:
          type: integer
          example: 2
        name:
          type: string
          example: Coffee
        description:
          type: string
          example: Restores focus for studying.
        price:
          type: integer
          example: 30
        imageUrl:
          type: string
          nullable: true
          example: /images/items/coffee.png
        iconEmoji:
          type: string
          nullable: true
          example: "\u2615"
        effectType:
          type: string
          nullable: true
          example: focus
    WalletResponse:
      type: object
      required: [coin]
      properties:
        coin:
          type: integer
          example: 150
    PurchaseRequest:
      type: object
      required: [itemId]
      properties:
        itemId:
          type: integer
          minimum: 1
          example: 2
        userId:
          type: integer
          minimum: 1
          description: 선택 필드입니다. 전달하면 토큰 사용자와 일치하는지 확인합니다.
          example: 1
    PurchaseResponse:
      type: object
      required: [coin, inventory]
      properties:
        coin:
          type: integer
          example: 120
        inventory:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
...
    InsufficientFundsResponse:
      type: object
      required: [error, coin]
      properties:
        error:
          type: string
          example: INSUFFICIENT_FUNDS
        coin:
          type: integer
          example: 10